feat: track migration statements in history table
