+++ internal/migration/apply/apply_test.go
@@ -1,23 +1,20 @@
 package apply
 
 import (
-	"bufio"
 	"context"
+	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgerrcode"
 	"github.com/spf13/afero"
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/supabase/cli/internal/migration/repair"
 	"github.com/supabase/cli/internal/testing/pgtest"
 	"github.com/supabase/cli/internal/utils"
-	"github.com/supabase/cli/internal/utils/parser"
 )
 
 func TestMigrateDatabase(t *testing.T) {
@@ -34,9 +31,11 @@ func TestMigrateDatabase(t *testing.T) {
 			Reply("CREATE SCHEMA").
 			Query(repair.CREATE_VERSION_TABLE).
 			Reply("CREATE TABLE").
+			Query(repair.ADD_STATEMENTS_COLUMN).
+			Reply("ALTER TABLE").
 			Query(sql).
 			Reply("CREATE SCHEMA").
-			Query(repair.INSERT_MIGRATION_VERSION, "0").
+			Query(repair.INSERT_MIGRATION_VERSION, "0", fmt.Sprintf("{%s}", sql)).
 			Reply("INSERT 1")
 		// Connect to mock
 		ctx := context.Background()
@@ -74,7 +73,8 @@ func TestMigrateUp(t *testing.T) {
 		conn.Query(repair.CREATE_VERSION_SCHEMA).
 			Reply("CREATE SCHEMA").
 			Query(repair.CREATE_VERSION_TABLE).
-			ReplyError(pgerrcode.InsufficientPrivilege, "permission denied for relation supabase_migrations")
+			ReplyError(pgerrcode.InsufficientPrivilege, "permission denied for relation supabase_migrations").
+			Query(repair.ADD_STATEMENTS_COLUMN)
 		// Connect to mock
 		ctx := context.Background()
 		mock, err := utils.ConnectLocalPostgres(ctx, pgconn.Config{Port: 5432}, conn.Intercept)
@@ -95,7 +95,9 @@ func TestMigrateUp(t *testing.T) {
 		conn.Query(repair.CREATE_VERSION_SCHEMA).
 			Reply("CREATE SCHEMA").
 			Query(repair.CREATE_VERSION_TABLE).
-			Reply("CREATE TABLE")
+			Reply("CREATE TABLE").
+			Query(repair.ADD_STATEMENTS_COLUMN).
+			Reply("ALTER TABLE")
 		// Connect to mock
 		ctx := context.Background()
 		mock, err := utils.ConnectLocalPostgres(ctx, pgconn.Config{Port: 5432}, conn.Intercept)
@@ -105,58 +107,5 @@ func TestMigrateUp(t *testing.T) {
 		err = MigrateUp(context.Background(), mock, []string{"20220727064247_missing.sql"}, fsys)
 		// Check error
 		assert.ErrorIs(t, err, os.ErrNotExist)
-	})
-}
-
-func TestMigrationFile(t *testing.T) {
-	t.Run("new from file sets max token", func(t *testing.T) {
-		viper.Reset()
-		// Setup in-memory fs
-		fsys := afero.NewMemMapFs()
-		// Setup initial migration
-		name := "20220727064247_create_table.sql"
-		path := filepath.Join(utils.MigrationsDir, name)
-		query := "BEGIN; " + strings.Repeat("a", parser.MaxScannerCapacity)
-		require.NoError(t, afero.WriteFile(fsys, path, []byte(query), 0644))
-		// Run test
-		migration, err := NewMigrationFromFile(path, fsys)
-		// Check error
-		assert.NoError(t, err)
-		assert.Len(t, migration.Lines, 2)
-		assert.Equal(t, "20220727064247", migration.version)
-	})
-
-	t.Run("new from reader errors on max token", func(t *testing.T) {
-		viper.Reset()
-		sql := "\tBEGIN; " + strings.Repeat("a", parser.MaxScannerCapacity)
-		// Run test
-		migration, err := NewMigrationFromReader(strings.NewReader(sql))
-		// Check error
-		assert.ErrorIs(t, err, bufio.ErrTooLong)
-		assert.ErrorContains(t, err, "After statement 1: \tBEGIN;")
-		assert.Nil(t, migration)
-	})
-
-	t.Run("throws error on insert failure", func(t *testing.T) {
-		migration := MigrationFile{
-			Lines:   []string{"create schema public"},
-			version: "0",
-		}
-		// Setup mock postgres
-		conn := pgtest.NewConn()
-		defer conn.Close(t)
-		conn.Query(migration.Lines[0]).
-			ReplyError(pgerrcode.DuplicateSchema, `schema "public" already exists`).
-			Query(repair.INSERT_MIGRATION_VERSION, "0")
-		// Connect to mock
-		ctx := context.Background()
-		mock, err := utils.ConnectLocalPostgres(ctx, pgconn.Config{Port: 5432}, conn.Intercept)
-		require.NoError(t, err)
-		defer mock.Close(ctx)
-		// Run test
-		err = migration.ExecBatch(context.Background(), mock)
-		// Check error
-		assert.ErrorContains(t, err, "ERROR: schema \"public\" already exists (SQLSTATE 42P06)")
-		assert.ErrorContains(t, err, "At statement 0: create schema public")
 	})
 }
